+++ client/src/components/SummaryItem.js
@@ -1,25 +1,24 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { makeStyles } from '@material-ui/styles';
 
 const useStyles = makeStyles({
-    image: {
-        padding: '7px',
-    },
-    icon: {
+    summaryItem: {
         display: 'flex',
         flexDirection: 'column',
-        justifyContent: 'center',
-        alignItems: 'center',
+        border: '1px solid #00000020',
         padding: '10px',
+        margin: '10px 0',
+        '& label': {
+            padding: '5px 0',
+        },
+        '& h2': {
+            padding: '5px 0',
+            margin: 0,
+        },
     },
 });
 
-export default function SummaryItem({ image, imageAlt, label }) {
+export default function SummaryItem({ children }) {
     const classes = useStyles();
-    return (
-        <div className={classes.icon}>
-            <img src={image} alt={imageAlt} />
-            <label>{label}</label>
-        </div>
-    );
+    return <div className={classes.summaryItem}>{children}</div>;
 }
